services/order-processor/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile '.\requirements.in'
#
kafka-python==2.3.0
    # via -r requirements.in
